+++ packages/at_client/test/notification_service_test.dart
@@ -3,6 +3,7 @@
 import 'dart:convert';
 
 import 'package:at_client/at_client.dart';
+import 'package:at_client/src/decryption_service/decryption_manager.dart';
 import 'package:at_client/src/encryption_service/encryption_manager.dart';
 import 'package:at_client/src/encryption_service/shared_key_encryption.dart';
 import 'package:at_client/src/manager/monitor.dart';
@@ -15,7 +16,8 @@ import 'package:at_persistence_secondary_server/at_persistence_secondary_server.
     hide AtNotification;
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
-import 'package:at_client/src/decryption_service/shared_key_decryption.dart';
+import 'package:at_client/src/response/at_notification.dart' as at_notification;
+import 'package:at_client/src/decryption_service/shared_with_me_decryption.dart';
 import 'package:uuid/uuid.dart';
 
 String? lastNotificationJson;
@@ -65,25 +67,26 @@ class FakeMonitor extends Fake implements Monitor {
   }
 }
 
+class MockAtKeyDecryptionManager extends Mock
+    implements AtKeyDecryptionManager {}
+
 // Mock class without implementation to throw exceptions
 class MockSharedKeyEncryption extends Mock implements SharedKeyEncryption {}
 
 // Mock class with implementation to populate metadata on encrypting value
 class MockSharedKeyEncryptionImpl extends Mock implements SharedKeyEncryption {
   @override
-  Future encrypt(AtKey atKey, value,
-      {bool storeSharedKeyEncryptedWithData = true}) async {
+  Future encrypt(AtKey atKey, value) async {
     //Set encryptionMetadata to atKey metadata
     atKey.metadata = Metadata();
-    if (storeSharedKeyEncryptedWithData) {
-      atKey.metadata.sharedKeyEnc = 'sharedKeyEnc';
-      atKey.metadata.pubKeyCS = 'publicKeyCS';
-    }
+    atKey.metadata.sharedKeyEnc = 'sharedKeyEnc';
+    atKey.metadata.pubKeyCS = 'publicKeyCS';
+
     return 'encryptedValue';
   }
 }
 
-class MockSharedKeyDecryption extends Mock implements SharedKeyDecryption {}
+class MockSharedKeyDecryption extends Mock implements SharedWithMeDecryption {}
 
 class MockAtClientManager extends Mock implements AtClientManager {}
 
@@ -104,13 +107,13 @@ class FakeAtKey extends Fake implements AtKey {}
 
 void main() {
   AtClientImpl mockAtClientImpl = MockAtClientImpl();
-  SharedKeyDecryption mockSharedKeyDecryption = MockSharedKeyDecryption();
+  AtKeyDecryptionManager mockDecryptionManager = MockAtKeyDecryptionManager();
+  SharedWithMeDecryption mockSharedKeyDecryption = MockSharedKeyDecryption();
   AtClientManager mockAtClientManager = MockAtClientManager();
   FakeMonitor fakeMonitor = FakeMonitor();
   SecondaryAddressFinder mockSecondaryAddressFinder =
       MockSecondaryAddressFinder();
-  AtKeyEncryptionManager mockAtKeyEncryptionManager =
-      MockAtKeyEncryptionManager();
+  AtKeyEncryptionManager mockEncryptionManager = MockAtKeyEncryptionManager();
   AtLookupImpl mockAtLookupImpl = MockAtLookupImpl();
   group('A group of test to validate notification request transformer', () {
     var value = '+91908909933';
@@ -498,6 +501,7 @@ void main() {
   group('A group of test to validate notification response transformer', () {
     setUp(() {
       registerFallbackValue(FakeAtKey());
+      mockDecryptionManager = MockAtKeyDecryptionManager();
       when(() => mockSharedKeyDecryption.decrypt(any(), 'encryptedValue'))
           .thenAnswer((_) => Future.value('decryptedValue'));
     });
@@ -559,16 +563,18 @@ void main() {
       bool? isEncrypted;
       var atNotification = AtNotification(
           '124',
-          'key-1',
+          '@bob:key-1.foo@alice',
           '@alice',
           '@bob',
           DateTime.now().millisecondsSinceEpoch,
           MessageTypeEnum.key.toString(),
           isEncrypted,
           value: 'encryptedValue');
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
@@ -585,16 +591,18 @@ void main() {
       bool isEncrypted = true;
       var atNotification = AtNotification(
           '124',
-          'key-1',
+          '@bob:key-1.foo@alice',
           '@alice',
           '@bob',
           DateTime.now().millisecondsSinceEpoch,
           MessageTypeEnum.key.toString(),
           isEncrypted,
           value: 'encryptedValue');
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
@@ -611,16 +619,18 @@ void main() {
       var isEncrypted = false;
       var atNotification = AtNotification(
           '124',
-          'key-1',
+          '@bob:key-1.foo@alice',
           '@alice',
           '@bob',
           DateTime.now().millisecondsSinceEpoch,
           MessageTypeEnum.key.toString(),
           isEncrypted,
           value: 'encryptedValue');
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
@@ -637,16 +647,18 @@ void main() {
       var isEncrypted = true;
       var atNotification = AtNotification(
           '124',
-          'key-1',
+          '@bob:key-1.foo@alice',
           '@alice',
           '@bob',
           DateTime.now().millisecondsSinceEpoch,
           MessageTypeEnum.key.toString(),
           isEncrypted,
           value: 'encryptedValue');
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
@@ -661,22 +673,24 @@ void main() {
       var isEncrypted = false;
       var atNotification = AtNotification(
           '124',
-          'key-1',
+          '@bob:key-1.foo@alice',
           '@alice',
           '@bob',
           DateTime.now().millisecondsSinceEpoch,
           MessageTypeEnum.key.toString(),
           isEncrypted);
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
             ..one = atNotification
             ..two = (NotificationConfig()..regex = '.*'));
       expect(transformedNotification.id, '124');
-      expect(transformedNotification.key, 'key-1');
+      expect(transformedNotification.key, '@bob:key-1.foo@alice');
     });
 
     test('A test to verify the reserved key is not decrypted', () async {
@@ -690,9 +704,11 @@ void main() {
           MessageTypeEnum.key.toString(),
           isEncrypted,
           value: 'encryptedValue');
-      var notificationResponseTransformer =
-          NotificationResponseTransformer(mockAtClientImpl);
-      notificationResponseTransformer.atKeyDecryption = mockSharedKeyDecryption;
+      var notificationResponseTransformer = NotificationResponseTransformer(
+          mockAtClientImpl,
+          decrypterManager: mockDecryptionManager);
+      when(() => mockDecryptionManager.get(any()))
+          .thenReturn(mockSharedKeyDecryption);
 
       var transformedNotification =
           await notificationResponseTransformer.transform(Tuple()
@@ -726,7 +742,7 @@ void main() {
           .thenAnswer((_) => Future.value(SecondaryAddress('dummyhost', 9001)));
       when(() => mockAtClientManager.secondaryAddressFinder)
           .thenAnswer((_) => mockSecondaryAddressFinder);
-      when(() => mockAtKeyEncryptionManager.get(atKey, currentAtSign))
+      when(() => mockEncryptionManager.get(atKey, currentAtSign))
           .thenAnswer((_) => mockSharedKeyEncryption);
       when(() => mockSharedKeyEncryption.encrypt(atKey, value)).thenThrow(
           AtPublicKeyNotFoundException(
@@ -743,8 +759,7 @@ void main() {
           atClientManager: mockAtClientManager,
           monitor: fakeMonitor) as NotificationServiceImpl;
 
-      notificationServiceImpl.atKeyEncryptionManager =
-          mockAtKeyEncryptionManager;
+      notificationServiceImpl.atKeyEncryptionManager = mockEncryptionManager;
 
       var notificationResult = await notificationServiceImpl.notify(
           NotificationParams.forUpdate(atKey, value: value),
@@ -770,7 +785,7 @@ void main() {
           .thenAnswer((_) => Future.value(SecondaryAddress('dummyhost', 9001)));
       when(() => mockAtClientManager.secondaryAddressFinder)
           .thenAnswer((_) => mockSecondaryAddressFinder);
-      when(() => mockAtKeyEncryptionManager.get(atKey, currentAtSign))
+      when(() => mockEncryptionManager.get(atKey, currentAtSign))
           .thenAnswer((_) => mockSharedKeyEncryption);
       when(() => mockSharedKeyEncryption.encrypt(atKey, value))
           .thenAnswer((_) => Future.value('encrypted_value'));
@@ -785,8 +800,7 @@ void main() {
           atClientManager: mockAtClientManager,
           monitor: fakeMonitor) as NotificationServiceImpl;
 
-      notificationServiceImpl.atKeyEncryptionManager =
-          mockAtKeyEncryptionManager;
+      notificationServiceImpl.atKeyEncryptionManager = mockEncryptionManager;
 
       var notificationResult = await notificationServiceImpl.notify(
           NotificationParams.forUpdate(atKey, value: value),