+++ tests/at_end2end_test/test/sharing_key_test.dart
@@ -1,7 +1,6 @@
 import 'dart:math';
 
 import 'package:at_client/at_client.dart';
-import 'package:at_client/src/encryption_service/encryption_manager.dart';
 import 'package:at_end2end_test/config/config_util.dart';
 import 'package:at_end2end_test/src/sync_initializer.dart';
 import 'package:at_end2end_test/src/test_initializers.dart';
@@ -11,12 +10,12 @@ import 'package:test/test.dart';
 import 'package:uuid/uuid.dart';
 
 void main() async {
-  late AtClientManager currentAtClientManager;
-  late AtClientManager sharedWithAtClientManager;
   late String currentAtSign;
   late String sharedWithAtSign;
   final namespace = TestConstants.namespace;
   var uuid = Uuid();
+
+  final acm = AtClientManager.getInstance();
 
   setUpAll(() async {
     currentAtSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
@@ -36,9 +35,8 @@ void main() async {
   test('Share a key to sharedWith atSign and lookup from sharedWith atSign',
       () async {
     // Setting currentAtSign atClient instance to context.
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+    await acm.setCurrentAtSign(currentAtSign, namespace,
+        TestPreferences.getInstance().getPreference(currentAtSign));
     // Generate  uuid
     var uniqueId = uuid.v4().hashCode;
     var phoneNumberKey = AtKey()
@@ -50,17 +48,14 @@ void main() async {
     // Appending a random number as a last number to generate a new phone number
     // for each run.
     var value = '+91 901920192${Random().nextInt(9)}';
-    var putResult =
-        await currentAtClientManager.atClient.put(phoneNumberKey, value);
+    var putResult = await acm.atClient.put(phoneNumberKey, value);
     expect(putResult, true);
-    await E2ESyncService.getInstance()
-        .syncData(currentAtClientManager.atClient.syncService);
+    await E2ESyncService.getInstance().syncData(acm.atClient.syncService);
 
     // Setting sharedWithAtSign atClient instance to context.
-    sharedWithAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(sharedWithAtSign, namespace,
-            TestPreferences.getInstance().getPreference(sharedWithAtSign));
-    var getResult = await sharedWithAtClientManager.atClient.get(AtKey()
+    await acm.setCurrentAtSign(sharedWithAtSign, namespace,
+        TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    var getResult = await acm.atClient.get(AtKey()
       ..key = 'phoneNumber-$uniqueId'
       ..sharedBy = currentAtSign);
     expect(getResult.value, value);
@@ -78,9 +73,8 @@ void main() async {
       'Create a key to sharedWith atSign with ttr and verify sharedWith atSign has a cached_key',
       () async {
     // Setting currentAtSign atClient instance to context.
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+    await acm.setCurrentAtSign(currentAtSign, namespace,
+        TestPreferences.getInstance().getPreference(currentAtSign));
     var uniqueId = uuid.v4().hashCode;
     var verificationKey = AtKey()
       ..key = 'verificationnumber-$uniqueId'
@@ -92,31 +86,27 @@ void main() async {
         ..ccd = true
         ..ttl = TestConstants.oneMinuteMillis);
     var value = '0873';
-    var putResult =
-        await currentAtClientManager.atClient.put(verificationKey, value);
+    var putResult = await acm.atClient.put(verificationKey, value);
     expect(putResult, true);
-    await E2ESyncService.getInstance()
-        .syncData(currentAtClientManager.atClient.syncService);
+    await E2ESyncService.getInstance().syncData(acm.atClient.syncService);
 
     // Setting sharedWithAtSign atClient instance to context.
-    sharedWithAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(sharedWithAtSign, namespace,
-            TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    await acm.setCurrentAtSign(sharedWithAtSign, namespace,
+        TestPreferences.getInstance().getPreference(sharedWithAtSign));
     var cachedVerificationKey = AtKey()
       ..key = 'verificationnumber-$uniqueId'
       ..sharedWith = sharedWithAtSign
       ..sharedBy = currentAtSign
       ..namespace = namespace
       ..metadata = (Metadata()..isCached = true);
-    await E2ESyncService.getInstance()
-        .syncData(sharedWithAtClientManager.atClient.syncService);
+    await E2ESyncService.getInstance().syncData(acm.atClient.syncService);
 
-    var getResult = await sharedWithAtClientManager.atClient
-        .getKeys(regex: cachedVerificationKey.toString());
+    var getResult =
+        await acm.atClient.getKeys(regex: cachedVerificationKey.toString());
     expect(getResult.contains(cachedVerificationKey.toString()), true);
 
-    AtValue getCachedKeyResponse = await sharedWithAtClientManager.atClient
-        .get(AtKey.fromString(getResult.first));
+    AtValue getCachedKeyResponse =
+        await acm.atClient.get(AtKey.fromString(getResult.first));
     expect(getCachedKeyResponse.value, '0873');
     expect(getCachedKeyResponse.metadata!.isCached, true);
   }, timeout: Timeout(Duration(minutes: 1)));
@@ -124,40 +114,24 @@ void main() async {
   /// The purpose of this test verify the following:
   /// 1. Backward compatibility for [metadata.sharedKeyEnc] and [metadata?.pubKeyCS]
   /// The encrypted value does not have new metadata but decrypt value successfully.
-  test(
-      'verify backward compatibility for sharedKey and checksum in metadata for shared key',
-      () async {
-    currentAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(currentAtSign, namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
+  test('Basic encrypted sharing test', () async {
     var uniqueId = uuid.v4().hashCode;
+    final value = 'New Jersey';
 
-    var locationKey = AtKey()
-      ..key = 'location-$uniqueId'
-      ..sharedWith = sharedWithAtSign
-      ..sharedBy = currentAtSign
-      ..namespace = namespace
-      ..metadata = Metadata();
+    await acm.setCurrentAtSign(currentAtSign, namespace,
+        TestPreferences.getInstance().getPreference(currentAtSign));
+    await acm.atClient.put(
+        AtKey()
+          ..key = 'location-$uniqueId'
+          ..sharedWith = sharedWithAtSign
+          ..sharedBy = currentAtSign
+          ..namespace = namespace,
+        value,
+        putRequestOptions: PutRequestOptions()..useRemoteAtServer = true);
 
-    var value = 'New Jersey';
-    var encryptionService =
-        AtKeyEncryptionManager(currentAtClientManager.atClient)
-            .get(locationKey, currentAtSign);
-    var encryptedValue = await encryptionService.encrypt(locationKey, value);
-    var result = await currentAtClientManager.atClient
-        .getRemoteSecondary()!
-        .executeCommand(
-            'update:ttl:300000:sharedKeyEnc:${locationKey.metadata.sharedKeyEnc}:pubKeyCS:${locationKey.metadata.pubKeyCS}:${locationKey.toString()} $encryptedValue\n',
-            auth: true);
-    expect(result != null, true);
-    await E2ESyncService.getInstance()
-        .syncData(currentAtClientManager.atClient.syncService);
-
-    sharedWithAtClientManager = await AtClientManager.getInstance()
-        .setCurrentAtSign(sharedWithAtSign, namespace,
-            TestPreferences.getInstance().getPreference(sharedWithAtSign));
-
-    var getResult = await sharedWithAtClientManager.atClient.get(AtKey()
+    await acm.setCurrentAtSign(sharedWithAtSign, namespace,
+        TestPreferences.getInstance().getPreference(sharedWithAtSign));
+    var getResult = await acm.atClient.get(AtKey()
       ..key = 'location-$uniqueId'
       ..sharedBy = currentAtSign);
     expect(getResult.value, value);